requirements.txt
Flask==3.0.3
pymupdf==1.24.10
openai-whisper==20250625
torch           # (optional, if you want GPU/CPU acceleration for whisper)
pymupdf
gunicorn